# Conflicts:
#	src/backend/ci/ext/tencent/remotedev/biz-remotedev-tencent/src/main/kotlin/com/tencent/devops/remotedev/service/expert/ExpertSupportService.kt
#	src/backend/ci/ext/tencent/remotedev/biz-remotedev-tencent/src/main/kotlin/com/tencent/devops/remotedev/service/gitproxy/GitProxyTGitService.kt

+++ src/backend/ci/ext/tencent/environment/biz-environment-tencent/src/main/kotlin/com/tencent/devops/environment/pojo/job/cmdbres/CmdbDataIns.kt
@@ -17,5 +17,5 @@ data class CmdbDataIns(
     @get:Schema(title = "服务器操作系统")
     val SfwName: String?,
     @get:Schema(title = "服务器ID")
-    val serverId: Int?
+    val serverId: Long?
 )
